defsystem.lisp: (already-loaded-systems)
defsystem.lisp: (already-loaded-systems): return the names of systems
which are known to be loaded. for now this means that they have an
entry in *MODULES* and a non nil COMPONENT-LOAD-TIME slot value.

Note that while component names are always lower case, the system name
preserves the case that the user supplied if the user supplied a
string

[240601]
move out mk::coerce-name into it's own commit.

[250128]
fixed typo in comments and commit message.