defsystem.lisp: (mklib): pathname production shenningans
defsystem.lisp: (mklib): add new keyword arguments no-lib-prefix,
no-library-suffix, and override, to control the name of the produced
target library. The initial choice was to generate and use
"lib<SYSTEM-NAME>-library.{so,a}" for shared and static targets and
"<SYSTEM-NAME>-library.<FASL>" for :fasl bundles, with a :defaults
providing the pathname defaults. However to be compatible with ecl
make-build we need to be able to omit the "lib prefix and "-library"
suffix, introduce options for those. Also add :override to skip the
whole naming mechanism and just use the given pathname.