defsystem.lisp: package-inferred-hack fixes 23Q4
* defsystem.lisp: (package-inferred-prefixp): extend to handle files
which define a package called a.b.c/d which uses a.b.c/e. This should
look for a file named "e.lisp" in the current directory. (hacky).
(make-mk-form): handle definitions that use :class
'asdf:package-inferred-system, fix "Could not infer the path"
diagnostic to print a newline. consider making it a warning.

[231218]
* defystem.lisp: (make-mk-form), some asds use a :class "STRING" where
"STRING" is some arbitrary subclass of
asdf:package-inferred-system. we have to edit those by hand anyway,
but we could fudge unknown classes as package-inferred systems here.

[240410]
* defsystem.lisp: (clean-components-list): make the cond an "econd".