defsystem.lisp: package-inferred-hack fixes 23Q4
* defsystem.lisp: (package-inferred-prefixp): extend to handle files
which define a package called a.b.c/d which uses a.b.c/e. This should
look for a file named "e.lisp" in the current directory. (hacky).
(make-mk-form): handle definitions that use :class
'asdf:package-inferred-system, fix "Could not infer the path"
diagnostic to print a newline. consider making it a warning.