defsystem.lisp: (oos): add a PROPAGATE keyword argument
* defsystem.lisp: (oos): add a PROPAGATE keyword argument which
binds *OPERATIONS-PROPAGATE-TO-SUBSYSTEMS*, so the current operation
to the current system. This was introduced in the argument list of
CLEAN-SYSTEM. One use case is to compile a system with FORCE under
sly/slime to navigate the errors in the compilation buffer.

[250328]
(compile-system): added :propagate keyword argument

[250305]
(load-system): added :propagate keyword argument.