defsystem.lisp: *dependencies-compilation-cache*
speed up "compile system" operation by skipping repeated "compile
system" on system dependencies

* defsystem.lisp: (*dependencies-compilation-cache*): new variable. If
set to T the speedup facility is disabled.
(operate-on-system): initialized here (if enabled).  Bound to a
hash-table which tracks systems that have been already compiled during
the current compile operation.  Keys are canonicalized system names
and the values are a cons of 2 numbers (ord . count) where ord
represents the order in which the system was compiled and count
represents the number of times a compile operation has been called on
that system.
(operate-on-component): updated here.
(operate-on-system-dependencies): skip the compile operation if the
system is found in the cache, while updating statistics.

statistics collection if needed has to be done manually by assigning a
hashtable and clearing it before each "compile system" operation.  By
default the statistics are discarded at the end of the operation.