squash! defsystem.lisp: *dependencies-compilation-cache*
defsystem.lisp: *compilation-dependencies-stats*

speed up "compile system" operation by avoiding repeated calls to
"compile system" on the system dependencies. (long standing bug)?

Optionally keep the information of the systems traversed during the
last mk-oos compile op in *last-compiled-system*
and *last-compiled-stats*, and retrieve it with (last-compiled-stats).

* defsystem.lisp: (*compilation-dependencies-stats*): new variable. If
set to T the speedup facility is disabled.
(operate-on-system): initialized here (if enabled).
(operate-on-component): updated here.
(operate-on-system-dependencies): skip the compile operation if the
system is found in the cache, while updating statistics.