defsystem.lisp: work around a limitation in the *compilation-dependencies-stats* implementation
when using the *compilation-dependencies-stats* hack to speed up
compile operations on a system, a limitation is that the hack requires
all depends-on deps to be already loaded. (Otherwise one hits an assert
which checks that the current operation is a compile operation)

This defect is mitigated by providing a restart that will load the
depends-on dependencies and restart the compile operation, and invoking
it automatically when the error is encountered.

* defsystem.lisp: (*compilation-dependencies-depth*): internal variable
to keep track of oos depth.
(with-compilation-dependencies-load-deps-handled): macro which
establishes a load-all-deps-and-retry restart around it's form.
(with-compilation-dependencies-load-deps-handled): macro which invokes
the load-all-deps-and-retry restart on error.  (operate-on-system):
use the first macro (operate-on-component): protect the assert with
the second macro.