defsystem.lisp: fix possible confusion about atom *central-registry*
*central-registry* could be a single location or a list of
locations. sds commit on 2025-08-16 introduced abstractions to deal
with *central-registry* that assumes they are always a list.

* defsystem.lisp: (add-registry-location): if *central-registry* is
not a list, promote it to a list before trying to add a new location.
(list-central-registry-directories): handle the case where
there *central-registry* is not a
list. (print-central-registry-directories):
use (list-central-registry-directories) instead of *central-registry*.

TODO: simplify other uses of *central-registry* that check for atom
vs. list to use list-central-registry-directories